requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# API requests
requests>=2.31.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Time series & statistics
statsmodels>=0.14.0
scipy>=1.11.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Geospatial
folium>=0.14.0
geopandas>=0.13.0
shapely>=2.0.0

# FlightRadarAPI (optional - educational wrapper)
FlightRadarAPI>=1.3.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Model explainability (optional)
shap>=0.42.0

# Progress bars
tqdm>=4.65.0
